+++ internal/lido/contracts/csaccounting/bonds_test.go
@@ -31,13 +31,16 @@ func TestBondSummary(t *testing.T) {
 		invalidID bool
 	}{
 		{
-			"BondSummary with valid ID, Holesky #1", "holesky", big.NewInt(2), false,
+			"BondSummary with valid ID, Mainnet", "mainnet", big.NewInt(2), false,
 		},
 		{
-			"BondSummary with valid ID, Holesky #2", "holesky", big.NewInt(14), false,
+			"BondSummary with valid ID, Holesky", "holesky", big.NewInt(14), false,
 		},
 		{
-			"BondSummary with invalid ID, Holesky ", "holesky", big.NewInt(-6), true,
+			"BondSummary with invalid ID, Holesky", "holesky", big.NewInt(-6), true,
+		},
+		{
+			"BondSummary with invalid ID, Mainnet", "mainnet", big.NewInt(-10), true,
 		},
 	}
 	for _, tc := range tcs {