+++ internal/pkg/clients/types.go
@@ -15,7 +15,9 @@ limitations under the License.
 */
 package clients
 
-import "github.com/NethermindEth/sedge/configs"
+import (
+	"github.com/NethermindEth/sedge/configs"
+)
 
 // Client : Struct Represent a client like geth, prysm, etc
 type Client struct {
@@ -34,6 +36,8 @@ func (c *Client) SetImageOrDefault(image string) {
 		c.setConsensusImage(image)
 	case "execution":
 		c.setExecutionImage(image)
+	case "distributedValidator":
+		c.setDistributedValidatorImage(image)
 	case "optimism":
 		c.setOptimismImage(image)
 	case "opexecution":
@@ -88,6 +92,15 @@ func (c *Client) setValidatorImage(image string) {
 	}
 }
 
+func (c *Client) setDistributedValidatorImage(image string) {
+	switch c.Name {
+	case "charon":
+		c.Image = valueOrDefault(image, configs.ClientImages.Distributed.Charon.String())
+	default:
+		c.Image = valueOrDefault(image, configs.ClientImages.Distributed.Charon.String())
+	}
+}
+
 func (c *Client) setOptimismImage(image string) {
 	switch c.Name {
 	case "opnode":
@@ -129,7 +142,7 @@ func valueOrDefault(value string, defaultValue string) string {
 	return value
 }
 
-// Clients : Struct Represent a combination of execution, consensus and validator clients
+// Clients : Struct Represent a combination of execution, consensus, validator and distributed validator clients
 type Clients struct {
 	Execution   *Client
 	Consensus   *Client
@@ -137,6 +150,7 @@ type Clients struct {
 	Optimism    *Client
 	Taiko       *Client
 	L2Execution *Client
+	DistributedValidator *Client
 }
 
 type ClientMap map[string]*Client