+++ CHANGELOG.md
@@ -8,12 +8,30 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- Added support for `op-geth` and `op-reth` clients on Optimism and Base networks.
+- New cli flag --distributed for running cluster with Charon distributed validator
+
+## [v1.7.2] - 2024-11-12
+
+### Fixed
+- Fix Erigon image source
+- Fix installation script path 
+
+## [v1.7.1] - 2024-11-1
+
+### Added
+- Updates on support for `op-geth` and `op-reth` clients on Optimism and Base networks.
 
 ### Changed
+- Updated Sedge's Docker commands internal functionality.
 - Changed the `--op-execution-image` flag to `--op-execution` on the `generate op-full-node` command.
 
+## [v1.7.0] - 2024-10-24
+
+### Added
+- Support Mainnet for Lido CSM setup.
+
 ## [v1.6.0] - 2024-10-18
+
 
 ### Added
 - New command `lido-status` to display data of Lido Node Operator.